+++ modal/_utils/blob_utils.py
@@ -188,16 +188,23 @@ def get_content_length(data: BinaryIO) -> int:
     return content_length - pos
 
 
-async def _blob_upload_with_fallback(items, blob_ids, callback):
+async def _blob_upload_with_fallback(items, blob_ids: list[str], callback) -> tuple[str, bool, int]:
+    r2_latency_ms = 0
+    r2_failed = False
     for idx, (item, blob_id) in enumerate(zip(items, blob_ids)):
         # We want to default to R2 95% of the time and S3 5% of the time.
         # To ensure the failure path is continuously exercised.
         if idx == 0 and len(items) > 1 and random.random() > HEALTHY_R2_UPLOAD_PERCENTAGE:
             continue
         try:
+            init_time = time.monotonic_ns()
             await callback(item)
-            return blob_id
+            if blob_id.endswith(":r2"):
+                r2_latency_ms = (time.monotonic_ns() - init_time) // 1_000_000
+            return blob_id, r2_failed, r2_latency_ms
         except Exception as _:
+            if blob_id.endswith(":r2"):
+                r2_failed = True
             # Ignore all errors except the last one, since we're out of fallback options.
             if idx == len(items) - 1:
                 raise
@@ -206,7 +213,7 @@ async def _blob_upload_with_fallback(items, blob_ids, callback):
 
 async def _blob_upload(
     upload_hashes: UploadHashes, data: Union[bytes, BinaryIO], stub, progress_report_cb: Optional[Callable] = None
-) -> str:
+) -> tuple[str, bool, int]:
     if isinstance(data, bytes):
         data = BytesIO(data)
 
@@ -232,7 +239,7 @@ async def _blob_upload(
                 progress_report_cb=progress_report_cb,
             )
 
-        blob_id = await _blob_upload_with_fallback(
+        blob_id, r2_failed, r2_latency_ms = await _blob_upload_with_fallback(
             resp.multiparts.items,
             resp.blob_ids,
             upload_multipart_upload,
@@ -252,7 +259,7 @@ async def _blob_upload(
                 content_md5_b64=upload_hashes.md5_base64,
             )
 
-        blob_id = await _blob_upload_with_fallback(
+        blob_id, r2_failed, r2_latency_ms = await _blob_upload_with_fallback(
             resp.upload_urls.items,
             resp.blob_ids,
             upload_to_s3_url,
@@ -261,10 +268,10 @@ async def _blob_upload(
     if progress_report_cb:
         progress_report_cb(complete=True)
 
-    return blob_id
+    return blob_id, r2_failed, r2_latency_ms
 
 
-async def blob_upload(payload: bytes, stub: ModalClientModal) -> str:
+async def blob_upload_with_r2_failure_info(payload: bytes, stub: ModalClientModal) -> tuple[str, bool, int]:
     size_mib = len(payload) / 1024 / 1024
     logger.debug(f"Uploading large blob of size {size_mib:.2f} MiB")
     t0 = time.time()
@@ -272,12 +279,17 @@ async def blob_upload(payload: bytes, stub: ModalClientModal) -> str:
         logger.warning("Blob uploading string, not bytes - auto-encoding as utf8")
         payload = payload.encode("utf8")
     upload_hashes = get_upload_hashes(payload)
-    blob_id = await _blob_upload(upload_hashes, payload, stub)
+    blob_id, r2_failed, r2_latency_ms = await _blob_upload(upload_hashes, payload, stub)
     dur_s = max(time.time() - t0, 0.001)  # avoid division by zero
     throughput_mib_s = (size_mib) / dur_s
     logger.debug(
         f"Uploaded large blob of size {size_mib:.2f} MiB ({throughput_mib_s:.2f} MiB/s, total {dur_s:.2f}s). {blob_id}"
     )
+    return blob_id, r2_failed, r2_latency_ms
+
+
+async def blob_upload(payload: bytes, stub: ModalClientModal) -> str:
+    blob_id, _, _ = await blob_upload_with_r2_failure_info(payload, stub)
     return blob_id
 
 
@@ -289,7 +301,8 @@ async def blob_upload_file(
     md5_hex: Optional[str] = None,
 ) -> str:
     upload_hashes = get_upload_hashes(file_obj, sha256_hex=sha256_hex, md5_hex=md5_hex)
-    return await _blob_upload(upload_hashes, file_obj, stub, progress_report_cb)
+    blob_id, _, _ = await _blob_upload(upload_hashes, file_obj, stub, progress_report_cb)
+    return blob_id
 
 
 @retry(n_attempts=5, base_delay=0.1, timeout=None)